고객관리 연결

+++ hotelServer/src/main/java/hms/hotelserver/ServerMain.java
@@ -43,6 +43,7 @@ public class ServerMain {
             ReservationService reservationService = new ReservationService();
             CheckOutService checkOutService = new CheckOutService();
             RoomAdminService roomService = new RoomAdminService();
+            CustomersService customerService = new CustomersService();
             StaffService staffService = new StaffService();
             
             String response = "ERROR|잘못된 요청";
@@ -51,37 +52,18 @@ public class ServerMain {
                 String[] parts = request.split("\\|");
                 String command = parts[0];
                 
-                //로그인
                 if (command.equals("LOGIN")) {
-                    // LOGIN|id|pw
-                    if (parts.length >= 3) {
-                        response = staffService.login(parts[1], parts[2]);
-                    } else {
-                        response = "ERROR|INVALID_FORMAT";
-                    }
+                    if (parts.length >= 3) { response = staffService.login(parts[1], parts[2]); }
+                    else { response = "ERROR|INVALID_FORMAT"; }
                  }
-                 
-                //직원 추가
                 else if (command.equals("STAFF_ADD")) {
-                    // STAFF_ADD|id|pw|role
-                    if (parts.length >= 4) {
-                        response = staffService.addStaff(parts[1], parts[2], parts[3]);
-                    } else {
-                        response = "ERROR|INVALID_FORMAT";
-                    }
+                    if (parts.length >= 4) {response = staffService.addStaff(parts[1], parts[2], parts[3]); }
+                    else { response = "ERROR|INVALID_FORMAT"; }
                 }
-                
-                //직원 삭제
                 else if (command.equals("STAFF_DELETE")) {
-                    // STAFF_DELETE|id
-                    if (parts.length >= 2) {
-                        response = staffService.deleteStaff(parts[1]);
-                    } else {
-                        response = "ERROR|INVALID_FORMAT";
-                    }
+                    if (parts.length >= 2) { response = staffService.deleteStaff(parts[1]); }
+                    else { response = "ERROR|INVALID_FORMAT"; }
                 }
-
-    
                 else if (command.equals("FIND_RESERVATION")) {
                     response = checkInService.findReservation(parts[1]);
                 }
@@ -102,6 +84,15 @@ public class ServerMain {
                 }
                 else if (command.equals("ROOMS_SERVICE_USAGE")){
                     response = checkOutService.loadRoomServices(Integer.parseInt(parts[1]));
+                }
+                else if(command.equals("SYNC_NEW_CUSTOMERS")){
+                    response = customerService.syncNewCustomers();
+                }
+                else if(command.equals("UPDATE_GRADE")){
+                    response = customerService.updateCustomerGrade(parts[1], parts[2], parts[3]);
+                }
+                else if(command.equals("GET_CUSTOMERS")){
+                    response = customerService.getAllCustomers();
                 }
                 
             }
